#120 prevent user from removing courses if they do not own the course in question

+++ backend/submit/remove-course.php
@@ -8,6 +8,16 @@ if (!$logged_in) {
   exit();
 }
 
+$b = false;
+$courses = $database->getCourses($user_id);
+for($x = 0; $x < sizeof($courses); $x++)
+  if($courses[$x]['id'] == $id)
+    $b = true;
+if(!$b) {
+  header('Location: ../../frontend/course.html?course=' . $id);
+  exit();
+}
+
 $database->removeCourse($course);
 header("Location: ../../frontend/index.html");
 exit();
